backend/requirements/base.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Environment
python-decouple==3.8
dj-database-url==2.1.0

# API Documentation
drf-spectacular==0.27.0

# Filtering
django-filter==23.5

# Validation
django-extensions==3.2.3

# Data Science & ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
prophet==1.1.4

# Utilities
python-dateutil==2.8.2
celery==5.3.4
